requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
autobahn[twisted]
twisted
numpy
soundfile
pydantic
python-multipart
scipy
requests
sounddevice
SpeechRecognition